feat: full-height layout with persistent right sidebar
