+++ src/components/Sidebar.module.css
@@ -5,9 +5,9 @@
   padding: 12px;
   background: rgba(255, 255, 255, 0.85);
   font-size: 14px;
-  width: 300px;
-  flex-shrink: 0;
-  overflow-y: auto;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
 }
 
 .filters {
@@ -66,5 +66,10 @@
 
 .flagLabel {
   font-size: 12px;
+}
+
+.bias {
+  margin-top: 8px;
+  display: block;
 }
