+++ src/pages/index.tsx
@@ -1,15 +1,23 @@
 import { TopBar } from '@/components/TopBar';
+import { SidebarLeft } from '@/components/SidebarLeft';
 import { SidebarRight } from '@/components/SidebarRight';
 import { HuGlobe } from '@/components/HuGlobe';
 
 export default function Home() {
   return (
-    <div style={{ display: 'flex', flexDirection: 'column', height: '100vh' }}>
-      <TopBar />
-      <div id="layout">
-        <main>
+    <div className="app">
+      <div className="topbar">
+        <TopBar />
+      </div>
+      <div className="left">
+        <SidebarLeft />
+      </div>
+      <div className="center">
+        <div className="globe">
           <HuGlobe />
-        </main>
+        </div>
+      </div>
+      <div className="right">
         <SidebarRight />
       </div>
     </div>
