+++ src/styles.css
@@ -10,7 +10,6 @@ html, body, #__next {
 
 body {
   font-family: sans-serif;
-  display: flex;
   overflow: hidden;
 }
 
@@ -20,13 +19,37 @@ body[data-theme='dark'] {
   --button-bg: #111;
   --button-color: #fff;
 }
-#layout {
-  display: flex;
-  flex: 1;
-  height: 100%;
+
+/* Layout grid */
+.app {
+  height: 100vh;
+  width: 100vw;
+  display: grid;
+  grid-template-rows: 56px 1fr;
+  grid-template-columns: 22% 1fr 28%;
 }
-main {
-  flex: 1;
-  position: relative;
-  min-width: 0;
+
+.topbar { grid-column: 1 / 4; grid-row: 1; }
+.left   { grid-column: 1; grid-row: 2; overflow: auto; }
+.center { grid-column: 2; grid-row: 2; position: relative; }
+.right  { grid-column: 3; grid-row: 2; overflow: auto; }
+
+.center > .globe {
+  position: absolute;
+  inset: 0;
 }
+
+@media (max-width: 1200px) {
+  .app { grid-template-columns: 20% 1fr 30%; }
+}
+
+@media (max-width: 900px) {
+  .app {
+    grid-template-columns: 100%;
+    grid-template-rows: 56px 1fr 40vh;
+  }
+  .left { display: none; }
+  .center { grid-row: 2; grid-column: 1; }
+  .right  { grid-row: 3; grid-column: 1; }
+}
+
